+++ src/components/country_iso_flag.jsx
@@ -1,0 +1,68 @@
+import worldwideImage from "../assets/images/worldwide.jpg";
+
+export default function getFlagOf(country) {
+  var countryISO = {
+    "Andorra":"AD",
+    "United Arab Emirates":"AE",
+    "Afghanistan":"AF",
+    "Antigua and Barbuda":"AG",
+    "Anguilla":"AG",
+    "Albania":"AL",
+    "Armenia":"AM",
+    "Netherlands Antilles":"AN",
+    "Angola":"AO",
+    "Antarctica":"AQ",
+    "Argentina":"AR",
+    "American Samoa":"AS",
+    "Austria":"AT",
+    "Australia":"AU",
+    "Aruba":"AW",
+    "Åland Islands":"AX",
+    "Azerbaijan":"AZ",
+    "Bosnia and Herzegovina":"BA",
+    "Barbados":"BB",
+    "Bangladesh":"BB",
+    "Belgium":"BE",
+    "Burkina Faso":"BF",
+    "Bulgaria":"BG",
+    "Bahrain":"BH",
+    "Burundi":"BI",
+    "Benin":"BJ",
+    "Croatia":"HR",
+    "Czech Republic":"CZ",
+    "Cyprus":"CY",
+    "Denmark":"DK",
+    "Estonia":"EE",
+    "Bermuda":"BM",
+    "France":"FR",
+    "French Guiana":"GF",
+    "Finland":"FI",
+    "Guadeloupe":"GE",
+    "Germany":"DE",
+    "Greece":"GR",
+    "Hungary":"HU",
+    "Ireland":"IE",
+    "Luxembourg":"LU",
+    "Lithuania":"LT",
+    "Latvia":"LV",
+    "Martinique":"MQ",
+    "Mayotte":"YT",
+    "Malta":"MT",
+    "Netherlands":"NL",
+    "Portugal":"PT",
+    "Poland":"PL",
+    "Réunion":"RE",
+    "Romania":"RO",
+    "Saint Barthélemy":"BL",
+    "Sweden":"SE",
+    "Spain":"ES",
+    "Saint Martin":"MF",
+    "Slovenia":"SL",
+    "Slovakia":"SK",
+    "Italy":"IT",
+  }
+  if(country === "Worldwide") {
+    return worldwideImage;
+  }
+  return `https://flagsapi.com/${countryISO[country]}/shiny/64.png`;
+}
